+++ protocol/signing/smime.go
@@ -1,4 +1,4 @@
-package signature
+package signing
 
 import (
 	"crypto/x509"
