+++ protocol/signing/ssh.go
@@ -1,4 +1,4 @@
-package signature
+package signing
 
 import (
 	"crypto/rand"
